Finish quest renaming and refactoring
