+++ inc/tasks/QuestDate.php
@@ -5,7 +5,9 @@
  * @package WapuuGotchi
  */
 
-namespace Wapuugotchi\Wapuugotchi;
+namespace Wapuugotchi\Wapuugotchi\Tasks;
+
+use Wapuugotchi\Wapuugotchi\Models\Quest;
 
 if ( ! defined( 'ABSPATH' ) ) :
 	exit();
@@ -32,11 +34,11 @@ class QuestDate {
 	 */
 	public function add_wapuugotchi_filter( $quests ) {
 		$default_quest = array(
-			new \Wapuugotchi\Wapuugotchi\Quest( 'login_1', null, __( 'Consistent Conquest: The Decennial Challenge', 'wapuugotchi' ), __( 'Prepare for an exciting adventure in the dynamic world of WordPress, where you\'re challenged to log in on 10 different days. Navigate through the digital landscape, ward off procrastination monsters, and claim the revered title of the Dedicated Regular. Are you ready to embrace the challenge and leave your mark in the vibrant realm of WordPress?', 'wapuugotchi' ), __( 'Nice, you logged in for 10 consecutive days!', 'wapuugotchi' ), 'success', 100, 1, 'Wapuugotchi\Wapuugotchi\QuestDate::always_true', 'Wapuugotchi\Wapuugotchi\QuestDate::login_completed_1' ),
-			new \Wapuugotchi\Wapuugotchi\Quest( 'login_2', 'login_1', __( 'Consistent Conquest: The Tenacity Trial', 'wapuugotchi' ), __( 'Prepare to test your persistence as you endeavor to log in for 20 days, facing increasingly formidable foes of distraction and embracing the mantle of the unwavering visitor.', 'wapuugotchi' ), __( 'Nice, you logged in for 20 consecutive days!', 'wapuugotchi' ), 'success', 100, 2, 'Wapuugotchi\Wapuugotchi\QuestDate::always_true', 'Wapuugotchi\Wapuugotchi\QuestDate::login_completed_2' ),
-			new \Wapuugotchi\Wapuugotchi\Quest( 'login_3', 'login_2', __( 'Consistent Conquest: The Decaday Triumph', 'wapuugotchi' ), __( 'The challenge continues as you navigate through the digital landscape, logging in for 30 consecutive days, proving your continued commitment and seizing victory as a faithful stalwart of the WordPress kingdom.', 'wapuugotchi' ), __( 'Nice, you logged in for 30 consecutive days!', 'wapuugotchi' ), 'success', 100, 3, 'Wapuugotchi\Wapuugotchi\QuestDate::always_true', 'Wapuugotchi\Wapuugotchi\QuestDate::login_completed_3' ),
-			new \Wapuugotchi\Wapuugotchi\Quest( 'login_4', 'login_3', __( 'Consistent Conquest: The 10-Day Crusade', 'wapuugotchi' ), __( 'Embark on this epic quest of regularity, where you must journey through the digital domain for 40 days and assert your unwavering presence as a valiant crusader dedicated to the WordPress realm.', 'wapuugotchi' ), __( 'Nice, you logged in for 40 consecutive days!', 'wapuugotchi' ), 'success', 100, 4, 'Wapuugotchi\Wapuugotchi\QuestDate::always_true', 'Wapuugotchi\Wapuugotchi\QuestDate::login_completed_4' ),
-			new \Wapuugotchi\Wapuugotchi\Quest( 'login_5', 'login_4', __( 'Consistent Conquest: The Perseverance Quest', 'wapuugotchi' ), __( 'Show your indomitable spirit in this quest, pushing the boundaries of dedication and logging in for 50 days, emerging as a champion of perseverance and a beacon of commitment in the vibrant realm of WordPress.', 'wapuugotchi' ), __( 'Nice, you logged in for 50 consecutive days!', 'wapuugotchi' ), 'success', 100, 5, 'Wapuugotchi\Wapuugotchi\QuestDate::always_true', 'Wapuugotchi\Wapuugotchi\QuestDate::login_completed_5' ),
+			new Quest( 'login_1', null, __( 'Consistent Conquest: The Decennial Challenge', 'wapuugotchi' ), __( 'Prepare for an exciting adventure in the dynamic world of WordPress, where you\'re challenged to log in on 10 different days. Navigate through the digital landscape, ward off procrastination monsters, and claim the revered title of the Dedicated Regular. Are you ready to embrace the challenge and leave your mark in the vibrant realm of WordPress?', 'wapuugotchi' ), __( 'Nice, you logged in for 10 consecutive days!', 'wapuugotchi' ), 'success', 100, 1, 'Wapuugotchi\Wapuugotchi\QuestDate::always_true', 'Wapuugotchi\Wapuugotchi\QuestDate::login_completed_1' ),
+			new Quest( 'login_2', 'login_1', __( 'Consistent Conquest: The Tenacity Trial', 'wapuugotchi' ), __( 'Prepare to test your persistence as you endeavor to log in for 20 days, facing increasingly formidable foes of distraction and embracing the mantle of the unwavering visitor.', 'wapuugotchi' ), __( 'Nice, you logged in for 20 consecutive days!', 'wapuugotchi' ), 'success', 100, 2, 'Wapuugotchi\Wapuugotchi\QuestDate::always_true', 'Wapuugotchi\Wapuugotchi\QuestDate::login_completed_2' ),
+			new Quest( 'login_3', 'login_2', __( 'Consistent Conquest: The Decaday Triumph', 'wapuugotchi' ), __( 'The challenge continues as you navigate through the digital landscape, logging in for 30 consecutive days, proving your continued commitment and seizing victory as a faithful stalwart of the WordPress kingdom.', 'wapuugotchi' ), __( 'Nice, you logged in for 30 consecutive days!', 'wapuugotchi' ), 'success', 100, 3, 'Wapuugotchi\Wapuugotchi\QuestDate::always_true', 'Wapuugotchi\Wapuugotchi\QuestDate::login_completed_3' ),
+			new Quest( 'login_4', 'login_3', __( 'Consistent Conquest: The 10-Day Crusade', 'wapuugotchi' ), __( 'Embark on this epic quest of regularity, where you must journey through the digital domain for 40 days and assert your unwavering presence as a valiant crusader dedicated to the WordPress realm.', 'wapuugotchi' ), __( 'Nice, you logged in for 40 consecutive days!', 'wapuugotchi' ), 'success', 100, 4, 'Wapuugotchi\Wapuugotchi\QuestDate::always_true', 'Wapuugotchi\Wapuugotchi\QuestDate::login_completed_4' ),
+			new Quest( 'login_5', 'login_4', __( 'Consistent Conquest: The Perseverance Quest', 'wapuugotchi' ), __( 'Show your indomitable spirit in this quest, pushing the boundaries of dedication and logging in for 50 days, emerging as a champion of perseverance and a beacon of commitment in the vibrant realm of WordPress.', 'wapuugotchi' ), __( 'Nice, you logged in for 50 consecutive days!', 'wapuugotchi' ), 'success', 100, 5, 'Wapuugotchi\Wapuugotchi\QuestDate::always_true', 'Wapuugotchi\Wapuugotchi\QuestDate::login_completed_5' ),
 		);
 
 		return array_merge( $default_quest, $quests );
