+++ inc/tasks/QuestPlugin.php
@@ -5,7 +5,9 @@
  * @package WapuuGotchi
  */
 
-namespace Wapuugotchi\Wapuugotchi;
+namespace Wapuugotchi\Wapuugotchi\Tasks;
+
+use Wapuugotchi\Wapuugotchi\Models\Quest;
 
 if ( ! defined( 'ABSPATH' ) ) :
 	exit();
@@ -32,9 +34,9 @@ class QuestPlugin {
 	 */
 	public function add_wapuugotchi_filter( $quests ) {
 		$default_quest = array(
-			new \Wapuugotchi\Wapuugotchi\Quest( 'use_seo_plugin_1', null, __( 'Search Spectacle', 'wapuugotchi' ), __( 'Get ready for an adventure in the WordPress realm, where you must equip your website with an SEO plugin to conquer the digital jungle. Navigate past treacherous search engine foes, outsmart the keyword dragons, and ascend to the coveted top spot in the search result kingdom. Are you prepared to heed the call and elevate your website to legendary optimization?', 'wapuugotchi' ), __( 'Awesome! &#128261;', 'wapuugotchi' ) . PHP_EOL . __( 'You\'ve installed a SEO plugin. Let\'s add some keywords so that search engines can find us better.', 'wapuugotchi' ), 'success', 100, 2, 'Wapuugotchi\Wapuugotchi\QuestPlugin::always_true', 'Wapuugotchi\Wapuugotchi\QuestPlugin::seo_installed_completed_1' ),
-			new \Wapuugotchi\Wapuugotchi\Quest( 'use_caching_plugin_1', null, __( 'Speedy Site Sprint', 'wapuugotchi' ), __( 'Embark on an exciting quest in the world of WordPress, where you must harness the power of a caching plugin to turbocharge your website`s performance. Navigate through the maze of lag, outsmart the slowdown gremlins, and unlock the potential for lightning-fast speed on your digital domain. Are you ready to rise to the challenge and propel your website to new heights of digital swiftness?', 'wapuugotchi' ), __( 'Fantastic! &#128171;', 'wapuugotchi' ) . PHP_EOL . __( 'You\'ve installed a caching plugin. This way we can increase the speed of our website.', 'wapuugotchi' ), 'success', 100, 2, 'Wapuugotchi\Wapuugotchi\QuestPlugin::always_true', 'Wapuugotchi\Wapuugotchi\QuestPlugin::caching_installed_completed_1' ),
-			new \Wapuugotchi\Wapuugotchi\Quest( 'use_security_plugin_1', null, __( 'Fortress Fortification', 'wapuugotchi' ), __( 'Venture into the realm of WordPress and heed the call to fortify your digital fortress. Arm your website with a formidable security plugin to ward off virtual villains and protect your online domain from the shadows of the web. Are you ready to take on the role of the valiant protector your website deserves?', 'wapuugotchi' ), __( 'Great!', 'wapuugotchi' ) . PHP_EOL . __( 'You\'ve installed a security plugin! This will help protecting me from villains. &#128170;', 'wapuugotchi' ), 'success', 100, 2, 'Wapuugotchi\Wapuugotchi\QuestPlugin::always_true', 'Wapuugotchi\Wapuugotchi\QuestPlugin::security_installed_completed_1' ),
+			new Quest( 'use_seo_plugin_1', null, __( 'Search Spectacle', 'wapuugotchi' ), __( 'Get ready for an adventure in the WordPress realm, where you must equip your website with an SEO plugin to conquer the digital jungle. Navigate past treacherous search engine foes, outsmart the keyword dragons, and ascend to the coveted top spot in the search result kingdom. Are you prepared to heed the call and elevate your website to legendary optimization?', 'wapuugotchi' ), __( 'Awesome! &#128261;', 'wapuugotchi' ) . PHP_EOL . __( 'You\'ve installed a SEO plugin. Let\'s add some keywords so that search engines can find us better.', 'wapuugotchi' ), 'success', 100, 2, 'Wapuugotchi\Wapuugotchi\QuestPlugin::always_true', 'Wapuugotchi\Wapuugotchi\QuestPlugin::seo_installed_completed_1' ),
+			new Quest( 'use_caching_plugin_1', null, __( 'Speedy Site Sprint', 'wapuugotchi' ), __( 'Embark on an exciting quest in the world of WordPress, where you must harness the power of a caching plugin to turbocharge your website`s performance. Navigate through the maze of lag, outsmart the slowdown gremlins, and unlock the potential for lightning-fast speed on your digital domain. Are you ready to rise to the challenge and propel your website to new heights of digital swiftness?', 'wapuugotchi' ), __( 'Fantastic! &#128171;', 'wapuugotchi' ) . PHP_EOL . __( 'You\'ve installed a caching plugin. This way we can increase the speed of our website.', 'wapuugotchi' ), 'success', 100, 2, 'Wapuugotchi\Wapuugotchi\QuestPlugin::always_true', 'Wapuugotchi\Wapuugotchi\QuestPlugin::caching_installed_completed_1' ),
+			new Quest( 'use_security_plugin_1', null, __( 'Fortress Fortification', 'wapuugotchi' ), __( 'Venture into the realm of WordPress and heed the call to fortify your digital fortress. Arm your website with a formidable security plugin to ward off virtual villains and protect your online domain from the shadows of the web. Are you ready to take on the role of the valiant protector your website deserves?', 'wapuugotchi' ), __( 'Great!', 'wapuugotchi' ) . PHP_EOL . __( 'You\'ve installed a security plugin! This will help protecting me from villains. &#128170;', 'wapuugotchi' ), 'success', 100, 2, 'Wapuugotchi\Wapuugotchi\QuestPlugin::always_true', 'Wapuugotchi\Wapuugotchi\QuestPlugin::security_installed_completed_1' ),
 		);
 
 		return array_merge( $default_quest, $quests );
