+++ inc/tasks/QuestStart.php
@@ -5,7 +5,9 @@
  * @package WapuuGotchi
  */
 
-namespace Wapuugotchi\Wapuugotchi;
+namespace Wapuugotchi\Tasks;
+
+use Wapuugotchi\Wapuugotchi\Models\Quest;
 
 if ( ! defined( 'ABSPATH' ) ) :
 	exit();
@@ -32,7 +34,7 @@ class QuestStart {
 	 */
 	public function add_wapuugotchi_filter( $quests ) {
 		$default_quest = array(
-			new \Wapuugotchi\Wapuugotchi\Quest( 'first_start_1', null, __( 'The Beginning', 'wapuugotchi' ), __( 'Embark into an amusing adventure in the WordPress realm with your Wapuu. Your noble endeavor giving him a home on your site, is the first step toward countless whimsical quests, unforeseen challenges, and surely a plethora of enthralling escapades. Are you ready to dive into the world of your WordPress and embrace the wild journey that lies ahead?', 'wapuugotchi' ), __( 'Thank you for giving me a home! &#10084;&#65039;', 'wapuugotchi' ), 'success', 100, 15, 'Wapuugotchi\Wapuugotchi\QuestStart::always_true', 'Wapuugotchi\Wapuugotchi\QuestStart::always_true' ),
+			new Quest( 'first_start_1', null, __( 'The Beginning', 'wapuugotchi' ), __( 'Embark into an amusing adventure in the WordPress realm with your Wapuu. Your noble endeavor giving him a home on your site, is the first step toward countless whimsical quests, unforeseen challenges, and surely a plethora of enthralling escapades. Are you ready to dive into the world of your WordPress and embrace the wild journey that lies ahead?', 'wapuugotchi' ), __( 'Thank you for giving me a home! &#10084;&#65039;', 'wapuugotchi' ), 'success', 100, 15, 'Wapuugotchi\Wapuugotchi\QuestStart::always_true', 'Wapuugotchi\Wapuugotchi\QuestStart::always_true' ),
 		);
 
 		return array_merge( $default_quest, $quests );
