+++ inc/tasks/QuestTheme.php
@@ -5,7 +5,9 @@
  * @package WapuuGotchi
  */
 
-namespace Wapuugotchi\Wapuugotchi;
+namespace Wapuugotchi\Tasks;
+
+use Wapuugotchi\Wapuugotchi\Models\Quest;
 
 if ( ! defined( 'ABSPATH' ) ) :
 	exit();
@@ -32,7 +34,7 @@ class QuestTheme {
 	 */
 	public function add_wapuugotchi_filter( $quests ) {
 		$default_quest = array(
-			new \Wapuugotchi\Wapuugotchi\Quest( 'cleanup_themes_1', null, __( 'Remove all unused themes', 'wapuugotchi' ), __( 'placeholder', 'wapuugotchi' ), __( 'You cleaned up! &#129529;', 'wapuugotchi' ) . PHP_EOL . 'We have only one theme now.', 'success', 100, 2, 'Wapuugotchi\Wapuugotchi\QuestTheme::always_true', 'Wapuugotchi\Wapuugotchi\QuestTheme::cleanup_themes_completed_1' ),
+			new Quest( 'cleanup_themes_1', null, __( 'Theme-tastic Tidiness', 'wapuugotchi' ), __( 'Embark on a daring adventure through the cluttered corridors of WordPress, where you must vanquish the unused themes that lurk in the shadows! Uninstall with precision, restore order, and emerge victorious as the Master of Theme Management!', 'wapuugotchi' ), __( 'You cleaned up! &#129529;', 'wapuugotchi' ) . PHP_EOL . 'We have only one theme now.', 'success', 100, 2, 'Wapuugotchi\Wapuugotchi\QuestTheme::always_true', 'Wapuugotchi\Wapuugotchi\QuestTheme::cleanup_themes_completed_1' ),
 		);
 
 		return array_merge( $default_quest, $quests );
